Consolidate mechanisms to render granule access links, so 508 changes can apply to both.
